+++ bathbot/src/active/impls/scores/user.rs
@@ -183,7 +183,7 @@ impl Display for OrderAppendix<'_> {
             }
             ScoresOrder::Bpm => {
                 let clock_rate = GameModsIntermode::from_bits(self.score.mods).legacy_clock_rate();
-                let bpm = self.map.map_or(0.0, |map| map.bpm) * clock_rate;
+                let bpm = self.map.map_or(0.0, |map| map.bpm) * clock_rate as f32;
 
                 write!(f, "`{}BPM`", round(bpm))
             }
@@ -205,7 +205,7 @@ impl Display for OrderAppendix<'_> {
             }
             ScoresOrder::Length => {
                 let clock_rate = GameModsIntermode::from_bits(self.score.mods).legacy_clock_rate();
-                let seconds_drain = self.map.map_or(0, |map| map.seconds_drain) as f32 / clock_rate;
+                let seconds_drain = self.map.map_or(0, |map| map.seconds_drain) as f64 / clock_rate;
 
                 write!(f, "`{}`", SecToMinSec::new(seconds_drain as u32))
             }
